chore: refactor new core package
stuff out of main and into core package
